requirement.txt
asgiref==3.4.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
django-crispy-forms==1.14.0
django-dotenv==1.4.2
gunicorn
psycopg2-binary
pint
django-extensions==3.1.3
django-widget-tweaks==1.4.8
10, 13, 15, 16
